Move VGSCheckoutUITestsFeature to separate file
